pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hermes"
description = "Mission analysis for HERON"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
version = "0.1.0"
dependencies = [
    "numpy",
    "celest==1.0.0",
    "supernova-cffi @ git+https://github.com/utat-ss/supernova",
]

[project.urls]
homepage = "https://github.com/utat-ss/hermes"
